# go_api/docs/swagger.yaml
definitions:
...
  dto.CourseDetails:
    properties:
      course_code:
        type: integer
      course_name:
        type: string
      ects_points:
        type: integer
    type: object
...
  dto.CreateCurriculumCourseRequest:
    properties:
      course:
        type: integer
      elective_block:
        type: integer
      major:
        type: integer
      semester:
        type: integer
      study_program:
        type: integer
    required:
    - course
    - semester
    - study_program
    type: object
  dto.CreateElectiveBlockRequest:
    properties:
      elective_block_name:
        type: string
      study_field:
        type: integer
    required:
    - elective_block_name
    - study_field
    type: object
  dto.CreateFacultyRequest:
    properties:
      faculty_name:
        type: string
      faculty_short:
        type: string
    required:
    - faculty_name
    - faculty_short
    type: object
  dto.CreateMajorRequest:
    properties:
      major_name:
        type: string
      study_field:
        type: integer
    required:
    - major_name
    - study_field
    type: object
...
  dto.CurriculumCourseResponse:
    properties:
      course:
        type: integer
      course_details:
        $ref: '#/definitions/dto.CourseDetails'
      elective_block:
        type: integer
      elective_block_details:
        $ref: '#/definitions/models.ElectiveBlock'
      major:
        type: integer
      major_details:
        $ref: '#/definitions/models.MajorReadResponse'
      semester:
        type: integer
      study_program:
        type: integer
    type: object
...
  dto.GroupResponse:
    properties:
      elective_block:
        type: integer
      group_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      major:
        type: integer
      semester:
        type: integer
      study_program:
        type: integer
    type: object
...
  models.Course:
    properties:
      course_code:
        type: integer
      course_coordinator:
        type: integer
      course_language:
        type: string
      course_name:
        type: string
      ects_points:
        type: integer
      leading_unit:
        type: integer
    type: object
...
  models.CurriculumCourse:
    properties:
      course:
        type: integer
      course_ref:
        allOf:
        - $ref: '#/definitions/models.Course'
        description: Relations
      elective_block:
        type: integer
      elective_block_ref:
        $ref: '#/definitions/models.ElectiveBlock'
      major:
        type: integer
      major_ref:
        $ref: '#/definitions/models.Major'
      semester:
        type: integer
      study_program:
        type: integer
      study_program_ref:
        $ref: '#/definitions/models.StudyProgram'
    type: object
  models.ElectiveBlock:
    properties:
      elective_block_name:
        type: string
      id:
        type: integer
      study_field:
        type: integer
    type: object
  models.Faculty:
    properties:
      faculty_name:
        type: string
      faculty_short:
        type: string
      id:
        type: integer
    type: object
  models.FacultyReadWithCounterResponse:
    properties:
      faculty_name:
        type: string
      faculty_short:
        type: string
      id:
        type: integer
      lecturers_count:
        type: integer
      students_count:
        type: integer
    type: object
  models.Group:
    properties:
      elective_block:
        type: integer
      group_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      major:
        type: integer
      members:
        items:
          $ref: '#/definitions/models.GroupMember'
        type: array
      semester:
        type: integer
      study_program:
        type: integer
    type: object
  models.GroupMember:
    properties:
      group:
        type: integer
      student:
        type: integer
    type: object
  models.Major:
    properties:
      id:
        type: integer
      major_name:
        type: string
      study_field:
        type: integer
    type: object
  models.MajorReadResponse:
    properties:
      group_count:
        type: integer
      id:
        type: integer
      major_name:
        type: string
      study_field:
        type: integer
    type: object
...
  models.PaginatedElectiveBlocksResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ElectiveBlock'
        type: array
    type: object
  models.PaginatedFacultiesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.FacultyReadWithCounterResponse'
        type: array
    type: object
  models.PaginatedMajorsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.MajorReadResponse'
        type: array
    type: object
...
  models.StudyProgram:
    properties:
      id:
        type: integer
      program_name:
        type: string
      start_year:
        type: string
      study_field:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/curriculum-courses:
    get:
      description: Returns curriculum courses with course, major and elective block
        details (with aggregated group counts)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/dto.CurriculumCourseResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get curriculum courses
      tags:
      - curriculum-courses
    post:
      consumes:
      - application/json
      description: Creates a new curriculum course entry.
      parameters:
      - description: Curriculum course data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCurriculumCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CurriculumCourse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create curriculum course
      tags:
      - curriculum-courses
  /api/v1/elective-blocks:
    get:
      description: Returns list of elective blocks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedElectiveBlocksResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get elective blocks
      tags:
      - elective-blocks
    post:
      consumes:
      - application/json
      description: Creates a new elective block.
      parameters:
      - description: Elective block data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateElectiveBlockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ElectiveBlock'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create elective block
      tags:
      - elective-blocks
  /api/v1/faculties:
    get:
      description: Returns faculties with lecturers and students count
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedFacultiesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get faculties
      tags:
      - faculties
    post:
      consumes:
      - application/json
      description: Creates a new faculty.
      parameters:
      - description: Faculty data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFacultyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Faculty'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create faculty
      tags:
      - faculties
  /api/v1/group-members:
    get:
      description: Returns list of group members with group relation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.GroupMember'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get group members
      tags:
      - group-members
    post:
      consumes:
      - application/json
      description: Adds a user to a group
      parameters:
      - description: Group member data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GroupMember'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/models.GroupMember'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create group member
      tags:
      - group-members
  /api/v1/groups:
    get:
      description: Returns list of all groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/dto.GroupResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new group.
      parameters:
      - description: Group data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/models.Group'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create group
      tags:
      - groups
  /api/v1/majors:
    get:
      description: Returns list of majors with group count
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedMajorsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get majors
      tags:
      - majors
    post:
      consumes:
      - application/json
      description: Creates a new major.
      parameters:
      - description: Major data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMajorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Major'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create major
      tags:
      - majors
...
swagger: "2.0"